Cleaned up formatting.
